+++ apps/app/src/features/search/client/components/SearchMethodMenuItem.tsx
@@ -1,35 +1,17 @@
 import React from 'react';
 
 import { useTranslation } from 'next-i18next';
-import { useRouter } from 'next/router';
-import { ListGroupItem } from 'reactstrap';
 
 import { useCurrentPagePath } from '~/stores/page';
 
-
-type MenuItemProps = {
-  children: React.ReactNode
-  href: string
-}
-
-const MenuItem = (props: MenuItemProps): JSX.Element => {
-  const { children, href } = props;
-  const router = useRouter();
-
-  return (
-    <ListGroupItem className="border-0 text-muted p-1 d-flex" tag="a" href={href} onClick={() => { router.push(href) }}>
-      <span className="material-symbols-outlined fs-4 me-3">search</span>
-      { children }
-    </ListGroupItem>
-  );
-};
+import { SearchMenuItem } from './SearchMenuItem';
 
 
-type SearchMethodMenuItemProps = {
+type Props = {
   searchKeyword: string
 }
 
-export const SearchMethodMenuItem = (props: SearchMethodMenuItemProps): JSX.Element => {
+export const SearchMethodMenuItem = (props: Props): JSX.Element => {
   const { t } = useTranslation('commons');
 
   const { data: currentPagePath } = useCurrentPagePath();
@@ -41,29 +23,32 @@ export const SearchMethodMenuItem = (props: SearchMethodMenuItemProps): JSX.Elem
   return (
     <div>
       { shouldShowButton && (
-        <MenuItem href={`/_search?q=${searchKeyword}`}>
+        <SearchMenuItem href={`/_search?q=${searchKeyword}`}>
+          <span className="material-symbols-outlined fs-4 me-3">search</span>
           <span>{searchKeyword}</span>
           <div className="ms-auto">
             <span>{t('search_method_menu_item.search_in_all')}</span>
           </div>
-        </MenuItem>
+        </SearchMenuItem>
       )}
 
-      <MenuItem href={`/_search?q=prefix:${currentPagePath} ${searchKeyword}`}>
+      <SearchMenuItem href={`/_search?q=prefix:${currentPagePath} ${searchKeyword}`}>
+        <span className="material-symbols-outlined fs-4 me-3">search</span>
         <code>prefix: {currentPagePath}</code>
         <span className="ms-2">{searchKeyword}</span>
         <div className="ms-auto">
           <span>{t('search_method_menu_item.only_children_of_this_tree')}</span>
         </div>
-      </MenuItem>
+      </SearchMenuItem>
 
       { shouldShowButton && (
-        <MenuItem href={`/_search?q="${searchKeyword}"`}>
+        <SearchMenuItem href={`/_search?q="${searchKeyword}"`}>
+          <span className="material-symbols-outlined fs-4 me-3">search</span>
           <span>{`"${searchKeyword}"`}</span>
           <div className="ms-auto">
             <span>{t('search_method_menu_item.exact_mutch')}</span>
           </div>
-        </MenuItem>
+        </SearchMenuItem>
       ) }
     </div>
   );
