+++ apps/app/src/features/search/client/components/SearchResultMenuItem.tsx
@@ -1,11 +1,11 @@
 import React from 'react';
 
 import { PagePathLabel, UserPicture } from '@growi/ui/dist/components';
-import { useRouter } from 'next/router';
-import { ListGroupItem } from 'reactstrap';
 import { useDebounce } from 'usehooks-ts';
 
 import { useSWRxSearch } from '~/stores/search';
+
+import { SearchMenuItem } from './SearchMenuItem';
 
 
 type Props = {
@@ -14,8 +14,6 @@ type Props = {
 
 export const SearchResultMenuItem = (props: Props): JSX.Element => {
   const { searchKeyword } = props;
-
-  const router = useRouter();
 
   const debouncedKeyword = useDebounce(searchKeyword, 500);
 
@@ -31,13 +29,7 @@ export const SearchResultMenuItem = (props: Props): JSX.Element => {
     <>
       <>
         {searchResult.data?.map(pageWithMeta => (
-          <ListGroupItem
-            key={pageWithMeta.data._id}
-            tag="a"
-            className="border-0 text-muted p-1 d-flex"
-            href={pageWithMeta.data._id}
-            onClick={() => { router.push(pageWithMeta.data._id) }}
-          >
+          <SearchMenuItem href={pageWithMeta.data._id} key={pageWithMeta.data._id}>
             <div className="mb-1 d-flex">
               <UserPicture user={pageWithMeta.data.creator} />
 
@@ -50,7 +42,7 @@ export const SearchResultMenuItem = (props: Props): JSX.Element => {
                 <span>{pageWithMeta.data.seenUsers.length}</span>
               </span>
             </div>
-          </ListGroupItem>
+          </SearchMenuItem>
         ))}
       </>
       <div className="border-top mb-2" />
